+++ cedar/auxiliaries/ParameterTemplate.fwd.h
@@ -1,6 +1,6 @@
 /*======================================================================================================================
 
-    Copyright 2011, 2012, 2013 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
+    Copyright 2011, 2012, 2013, 2014 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
  
     This file is part of cedar.
 
@@ -53,7 +53,7 @@ namespace cedar
 {
   namespace aux
   {
-    template <typename T> class ParameterTemplate;
+    template <typename T, typename ValuePolicy> class ParameterTemplate;
   }
 }