+++ cedar/auxiliaries/gui/Configurable.h
@@ -1,6 +1,6 @@
 /*======================================================================================================================
 
-    Copyright 2011, 2012 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
+    Copyright 2011, 2012, 2013, 2014 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
  
     This file is part of cedar.
 
@@ -55,6 +55,8 @@
 #endif // Q_MOC_RUN
 #include <QWidget>
 #include <QTreeWidget>
+#include <map>
+#include <string>
 
 
 /*!@brief A widget to display and manipulate the parameters of cedar::aux::Configurables.
@@ -98,6 +100,9 @@ public:
 
 public slots:
   void fitRowsToContents();
+
+signals:
+  void settingsChanged();
 
   //--------------------------------------------------------------------------------------------------------------------
   // protected methods