+++ cedar/processing/gui/Ide.cpp
@@ -1,6 +1,6 @@
 /*======================================================================================================================
 
-    Copyright 2011, 2012, 2013 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
+    Copyright 2011, 2012, 2013, 2014 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
  
     This file is part of cedar.
 
@@ -89,7 +89,8 @@ cedar::proc::gui::Ide::Ide(bool loadDefaultPlugins, bool redirectLogToGui)
 mpConsistencyChecker(NULL),
 mpPerformanceOverview(NULL),
 mpConsistencyDock(NULL),
-mpBoostControl(NULL)
+mpBoostControl(NULL),
+mSuppressCloseDialog(false)
 {
   // setup the (automatically generated) ui components
   this->setupUi(this);
@@ -256,6 +257,22 @@ mpBoostControl(NULL)
 
   QObject::connect(mpActionPerformanceOverview, SIGNAL(triggered()), this->mpPerformanceOverview, SLOT(show()));
 
+
+  QObject::connect(this->mpRecorderWidget,
+                   SIGNAL(settingsChanged()),
+                   this,
+                   SLOT(architectureChanged()));
+
+  QObject::connect(this->mpPropertyTable,
+                   SIGNAL(settingsChanged()),
+                   this,
+                   SLOT(architectureChanged()));
+
+  QObject::connect(this->mpProcessingDrawer->getScene(),
+                   SIGNAL(sceneChanged()),
+                   this,
+                   SLOT(architectureChanged()));
+
   cedar::aux::PluginProxy::connectToPluginDeclaredSignal
   (
     boost::bind(&cedar::proc::gui::Ide::resetStepList, this)
@@ -275,6 +292,17 @@ cedar::proc::gui::Ide::~Ide()
 //----------------------------------------------------------------------------------------------------------------------
 // methods
 //----------------------------------------------------------------------------------------------------------------------
+
+void cedar::proc::gui::Ide::setArchitectureChanged(bool changed)
+{
+  this->mpActionSave->setEnabled(changed);
+  this->setWindowModified(changed);
+}
+
+void cedar::proc::gui::Ide::architectureChanged()
+{
+  this->setArchitectureChanged(true);
+}
 
 void cedar::proc::gui::Ide::globalTimeFactorSliderChanged(int newValue)
 {
@@ -302,7 +330,7 @@ void cedar::proc::gui::Ide::showRobotManager()
 
 void cedar::proc::gui::Ide::displayFilename(const std::string& filename)
 {
-  this->setWindowTitle(this->mDefaultWindowTitle + " - " + QString::fromStdString(filename));
+  this->setWindowTitle(this->mDefaultWindowTitle + " - " + QString::fromStdString(filename) + "[*]");
 }
 
 void cedar::proc::gui::Ide::showBoostControl()
@@ -463,7 +491,7 @@ void cedar::proc::gui::Ide::showAboutDialog()
   QImage version_image(":/cedar/processing/gui/images/current_version_image.svg");
   p_version_image->setPixmap(QPixmap::fromImage(version_image));
 
-  QString about_text = "<center>This is cedar's processingIde<br />built with<br />cedar version <b>";
+  QString about_text = "<center>This is cedar<br />built with library<br />version <b>";
   about_text += QString::fromStdString(cedar::aux::versionNumberToString(CEDAR_VERSION));
   about_text += "</b>"
 #ifdef DEBUG
@@ -510,10 +538,54 @@ void cedar::proc::gui::Ide::toggleGrid(bool triggered)
   }
 }
 
+bool cedar::proc::gui::Ide::checkSave()
+{
+  if (this->isWindowModified() && !mSuppressCloseDialog)
+  {
+    auto r = QMessageBox::question
+        (
+          this,
+          "Save changes?",
+          "There are unsaved changes. Do you want to close them?",
+          QMessageBox::Save | QMessageBox::Cancel | QMessageBox::Discard
+        );
+
+    switch (r)
+    {
+      case QMessageBox::Save:
+        // let's not accept the event yet, in case anything goes wrong (like an exception)
+        return this->save();
+
+      case QMessageBox::Cancel:
+        return false;
+
+      case QMessageBox::Discard:
+        // nothing to do: discard changes
+        break;
+
+      default:
+        CEDAR_ASSERT(false);
+    }
+  }
+  return true;
+}
+
+
 void cedar::proc::gui::Ide::closeEvent(QCloseEvent *pEvent)
 {
+  // let's not accept the event yet, in case anything goes wrong (like an exception)
+  pEvent->setAccepted(false);
+  if (this->checkSave())
+  {
+    pEvent->setAccepted(true);
+  }
+  else
+  {
+    return;
+  }
+
   this->storeSettings();
-  // Without this, the gui_ProcessingIde crashes when exiting in certain circumstances (see unit test gui_ProcessingIde)
+  // Without this, the GUI crashes when exiting in certain circumstances (see unit test gui_cedar)
   this->mpPropertyTable->clear();
   pEvent->accept();
 }
@@ -797,7 +869,11 @@ void cedar::proc::gui::Ide::stopThreads()
 
 void cedar::proc::gui::Ide::newFile()
 {
-  this->mpActionSave->setEnabled(false);
+  if (!this->checkSave())
+  {
+    return;
+  }
+
   this->resetTo(cedar::proc::gui::NetworkPtr(new cedar::proc::gui::Network(this, this->mpProcessingDrawer->getScene())));
 
   this->displayFilename("unnamed file");
@@ -806,15 +882,26 @@ void cedar::proc::gui::Ide::newFile()
   this->mpActionToggleSmartConnections->blockSignals(true);
   this->mpActionToggleSmartConnections->setChecked(this->mNetwork->getSmartConnection());
   this->mpActionToggleSmartConnections->blockSignals(false);
+
+  this->setArchitectureChanged(false);
 }
 
-void cedar::proc::gui::Ide::save()
+bool cedar::proc::gui::Ide::save()
 {
-  this->mNetwork->write();
-  cedar::proc::gui::SettingsSingleton::getInstance()->appendArchitectureFileToHistory(this->mNetwork->getFileName());
+  if (this->mNetwork->getFileName().empty())
+  {
+    return this->saveAs();
+  }
+  else
+  {
+    this->mNetwork->write();
+    cedar::proc::gui::SettingsSingleton::getInstance()->appendArchitectureFileToHistory(this->mNetwork->getFileName());
+    this->setArchitectureChanged(false);
+    return true;
+  }
 }
 
-void cedar::proc::gui::Ide::saveAs()
+bool cedar::proc::gui::Ide::saveAs()
 {
   cedar::aux::DirectoryParameterPtr last_dir = cedar::proc::gui::SettingsSingleton::getInstance()->lastArchitectureLoadDialogDirectory();
 
@@ -824,30 +911,37 @@ void cedar::proc::gui::Ide::saveAs()
                                               "json (*.json)" // filter(s), separated by ';;'
                                               );
 
-  if (!file.isEmpty())
+  if (file.isEmpty())
   {
-    if (!file.endsWith(".json"))
-    {
-      file += ".json";
-    }
-
-    this->mNetwork->write(file.toStdString());
-    this->displayFilename(file.toStdString());
-
-    cedar::proc::gui::SettingsSingleton::getInstance()->appendArchitectureFileToHistory(file.toStdString());
-
-    this->mpActionSave->setEnabled(true);
-
-    cedar::proc::gui::SettingsSingleton::getInstance()->appendArchitectureFileToHistory(file.toStdString());
-    QString path = file.remove(file.lastIndexOf(QDir::separator()), file.length());
-    cedar::aux::DirectoryParameterPtr last_dir = cedar::proc::gui::SettingsSingleton::getInstance()->lastArchitectureLoadDialogDirectory();
-    last_dir->setValue(path);
-
+    return false;
   }
+
+  if (!file.endsWith(".json"))
+  {
+    file += ".json";
+  }
+
+  this->mNetwork->write(file.toStdString());
+  this->displayFilename(file.toStdString());
+  this->setArchitectureChanged(false);
+
+  cedar::proc::gui::SettingsSingleton::getInstance()->appendArchitectureFileToHistory(file.toStdString());
+
+  this->mpActionSave->setEnabled(true);
+
+  QString path = file.remove(file.lastIndexOf(QDir::separator()), file.length());
+  last_dir->setValue(path);
+
+  return true;
 }
 
 void cedar::proc::gui::Ide::load()
 {
+  if (!this->checkSave())
+  {
+    return;
+  }
+
   cedar::aux::DirectoryParameterPtr last_dir = cedar::proc::gui::SettingsSingleton::getInstance()->lastArchitectureLoadDialogDirectory();
 
   QString file = QFileDialog::getOpenFileName(this, // parent
@@ -864,6 +958,11 @@ void cedar::proc::gui::Ide::load()
 
 void cedar::proc::gui::Ide::loadFile(QString file)
 {
+  if (!this->checkSave())
+  {
+    return;
+  }
+
   // print message
   cedar::aux::LogSingleton::getInstance()->message
                                            (
@@ -1015,7 +1114,6 @@ void cedar::proc::gui::Ide::loadFile(QString file)
     p_dialog->displayStdException(e);
     p_dialog->exec();
   }
-  this->mpActionSave->setEnabled(true);
 
   this->setNetwork(network);
 
@@ -1030,6 +1128,8 @@ void cedar::proc::gui::Ide::loadFile(QString file)
   this->mpActionToggleSmartConnections->blockSignals(true);
   this->mpActionToggleSmartConnections->setChecked(this->mNetwork->getSmartConnection());
   this->mpActionToggleSmartConnections->blockSignals(false);
+
+  this->setArchitectureChanged(false);
 }
 
 void cedar::proc::gui::Ide::keyPressEvent(QKeyEvent* pEvent)
@@ -1115,6 +1215,8 @@ void cedar::proc::gui::Ide::showTriggerConnections(bool show)
 void cedar::proc::gui::Ide::toggleSmartConnections(bool smart)
 {
   this->mNetwork->toggleSmartConnectionMode(smart);
+
+  this->setArchitectureChanged(true);
 }
 
 void cedar::proc::gui::Ide::closePlots()
@@ -1165,6 +1267,8 @@ void cedar::proc::gui::Ide::addPlotGroup()
     int pos = this->mpPlotGroupsComboBox->count();
     this->mpPlotGroupsComboBox->insertItem(pos, text);
     this->mpPlotGroupsComboBox->setCurrentIndex(pos);
+
+    this->setArchitectureChanged(true);
   }
 }
   
@@ -1183,6 +1287,8 @@ void cedar::proc::gui::Ide::editPlotGroup()
       this->mpPlotGroupsComboBox->removeItem(position);
       this->mpPlotGroupsComboBox->insertItem(position, text);
       this->mpPlotGroupsComboBox->setCurrentIndex(position);
+
+      this->setArchitectureChanged(true);
     }
   }
 }
@@ -1193,6 +1299,8 @@ void cedar::proc::gui::Ide::displayPlotGroup()
   if(this->mpPlotGroupsComboBox->currentIndex() != -1)
   {
     this->mNetwork->displayPlotGroup(plot_group_name.toStdString()); // toStdString assumes ascii
+
+    this->setArchitectureChanged(true);
   }
 }
 
@@ -1204,11 +1312,13 @@ void cedar::proc::gui::Ide::deletePlotGroup()
   {
     QMessageBox::StandardButton reply;
     QString message = QString("This will delete %1. Proceed?").arg(plot_group_name);
-    reply = QMessageBox::question(this, "Delete Plot Group", message, QMessageBox::Yes|QMessageBox::No);
+    reply = QMessageBox::question(this, "Delete Plot Group", message, QMessageBox::Yes | QMessageBox::No);
     if (reply == QMessageBox::Yes)
     {
       this->mNetwork->removePlotGroup(plot_group_name.toStdString()); // toStdString assumes ascii
       this->mpPlotGroupsComboBox->removeItem(position);
+
+      this->setArchitectureChanged(true);
     }
   }
 }