+++ executables/cedar-shell/cedar-shell.cpp
@@ -1,7 +1,7 @@
 /*======================================================================================================================
 
-    Copyright 2011, 2012, 2013 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
- 
+    Copyright 2011, 2012, 2013, 2014 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
+
     This file is part of cedar.
 
     cedar is free software: you can redistribute it and/or modify it under
@@ -22,42 +22,39 @@
     Institute:   Ruhr-Universitaet Bochum
                  Institut fuer Neuroinformatik
 
-    File:        MainApplication.fwd.h
+    File:        cedar-shell.cpp
 
     Maintainer:  Oliver Lomp
+
     Email:       oliver.lomp@ini.ruhr-uni-bochum.de
+
     Date:        2014 01 20
 
-    Description: Forward declaration file for the class cedar::processingCL::MainApplication.
+    Description: A command-line version of cedar.
 
     Credits:
 
 ======================================================================================================================*/
 
-#ifndef CEDAR_PROC_CL_MAIN_APPLICATION_FWD_H
-#define CEDAR_PROC_CL_MAIN_APPLICATION_FWD_H
-
-// CEDAR CONFIGURATION
-#include "cedar/configuration.h"
-#include "cedar/defines.h"
-
 // CEDAR INCLUDES
 
+// LOCAL INCLUDES
+#include "MainApplication.h"
+
 // SYSTEM INCLUDES
-#ifndef Q_MOC_RUN
-  #include <boost/smart_ptr.hpp>
-#endif // Q_MOC_RUN
+#include <QApplication>
+#include <QTimer>
 
-//!@cond SKIPPED_DOCUMENTATION
-namespace cedar
+
+int main(int argc, char** argv)
 {
-  namespace processingCL
-  {
-    class MainApplication;
-    CEDAR_GENERATE_POINTER_TYPES(MainApplication);
-  }
+  QApplication app(argc, argv);
+
+  cedar::processingCL::MainApplicationPtr application(new cedar::processingCL::MainApplication(argc, argv));
+
+  QObject::connect(application.get(), SIGNAL(quit()), &app, SLOT(quit()));
+
+  QTimer::singleShot(0, application.get(), SLOT(exec()));
+
+  return app.exec();
 }
-//!@endcond
-
-#endif // CEDAR_PROC_CL_MAIN_APPLICATION_FWD_H
-