+++ tests/unit/processing/gui/gui_cedar/gui_cedar.cpp
@@ -1,6 +1,6 @@
 /*======================================================================================================================
 
-    Copyright 2011, 2012, 2013 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
+    Copyright 2011, 2012, 2013, 2014 Institut fuer Neuroinformatik, Ruhr-Universitaet Bochum, Germany
 
     This file is part of cedar.
 
@@ -71,6 +71,7 @@ int main(int argc, char** argv)
   int errors = 0;
 
   cedar::proc::gui::Ide* p_ide = new cedar::proc::gui::Ide(false);
+  p_ide->suppressCloseDialog(true);
   {
     simulateEventLoop();